Restructure components
